fix(suite): divide metadata actions to thunks and plain actions to prevent circular esm
