+++ packages/suite/src/utils/suite/metadata.ts
@@ -4,7 +4,7 @@ import * as crypto from 'crypto';
 import { DataType, MetadataProvider } from '@suite-common/metadata-types';
 import { StaticSessionId } from '@trezor/connect';
 
-import { FetchIntervalTrackingId } from 'src/actions/suite/metadataProviderActions';
+import { FetchIntervalTrackingId } from 'src/actions/suite/metadataProviderThunks';
 
 // note we only need base58 conversion fn from base58check, other functions from there might
 // be supplemented from crypto module
