+++ packages/suite/src/views/wallet/details/index.tsx
@@ -13,7 +13,7 @@ import {
     Url,
 } from '@trezor/urls';
 
-import { exportMetadataToBip329File } from 'src/actions/suite/metadataActions';
+import { exportMetadataToBip329File } from 'src/actions/suite/metadataThunks';
 import { showXpub } from 'src/actions/wallet/publicKeyActions';
 import { AccountTypeBadge } from 'src/components/suite/AccountTypeBadge';
 import { LearnMoreButton } from 'src/components/suite/LearnMoreButton';
